모집글, 스크랩 리스트 조회 기능 추가, 수정

+++ src/main/java/com/justcodeit/moyeo/study/application/scrap/ScrapService.java
@@ -1,8 +1,15 @@
 package com.justcodeit.moyeo.study.application.scrap;
 
+import com.justcodeit.moyeo.study.application.post.exception.PostCannotFoundException;
+import com.justcodeit.moyeo.study.application.scrap.exception.PostAlreadyDeletedException;
+import com.justcodeit.moyeo.study.application.scrap.exception.PostAlreadyScrappedException;
 import com.justcodeit.moyeo.study.application.scrap.exception.ScrapCannotFoundException;
-import com.justcodeit.moyeo.study.interfaces.dto.scrap.ScrapQueryDto;
+import com.justcodeit.moyeo.study.application.scrap.exception.ScrapNotAuthorizedException;
+import com.justcodeit.moyeo.study.model.inquiry.ScrapQueryDto;
+import com.justcodeit.moyeo.study.model.post.PostStatus;
+import com.justcodeit.moyeo.study.persistence.Post;
 import com.justcodeit.moyeo.study.persistence.Scrap;
+import com.justcodeit.moyeo.study.persistence.repository.PostRepository;
 import com.justcodeit.moyeo.study.persistence.repository.scrap.ScrapRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -15,17 +22,29 @@ import java.util.List;
 public class ScrapService {
 
   private final ScrapRepository scrapRepository;
+  private final PostRepository postRepository;
 
-  // TODO : post 가 이미 삭제된 상태에 대한 exception 추가 예정
   @Transactional
   public Long makeScrap(String userId, Long postId) {
-    Scrap scrap = new Scrap(userId, postId);
+    Post post = postRepository.findById(postId)
+            .orElseThrow(PostCannotFoundException::new);
+
+    validatePostByUserId(userId, post);
+    Scrap scrap = new Scrap(userId, post.getId());
+
     return scrapRepository.save(scrap).getId();
   }
 
   @Transactional
-  public void deleteScrap(Long scrapId) {
-    Scrap scrap = findScrap(scrapId);
+  public void deleteScrap(String userId, Long postId) {
+    Post post = postRepository.findById(postId)
+            .orElseThrow(PostCannotFoundException::new);
+    Scrap scrap = scrapRepository.findByUserIdAndPostId(userId, post.getId())
+            .orElseThrow(ScrapCannotFoundException::new);
+
+    if (!scrap.getUserId().equals(userId)) {
+      throw new ScrapNotAuthorizedException();
+    }
     scrapRepository.delete(scrap);
   }
 
@@ -34,9 +53,13 @@ public class ScrapService {
     return scrapRepository.findScrapListByUserId(userId);
   }
 
-  @Transactional(readOnly = true)
-  private Scrap findScrap(Long scrapId) {
-    return scrapRepository.findById(scrapId)
-            .orElseThrow(() -> new ScrapCannotFoundException(String.format("해당 스크랩을 찾을 수 없습니다 : %s", scrapId)));
+  private void validatePostByUserId(String userId, Post post) {
+    if (post.getPostStatus() != PostStatus.NORMAL) {
+      throw new PostAlreadyDeletedException("이미 삭제된 모집글입니다");
+    }
+
+    if (scrapRepository.existsByUserIdAndPostId(userId, post.getId())) {
+      throw new PostAlreadyScrappedException("이미 스크랩이 완료된 모집글입니다");
+    }
   }
 }
